rename inode_deviceid to natice && remove support for inode_marker on windows
